+++ lib/traces/provider/async/service/supervisor.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+# Released under the MIT License.
+# Copyright, 2026, by Samuel Williams.
+
+require_relative "supervisor/server"
+require_relative "supervisor/supervisor_controller"